requirements.txt
# Lane Google Requirements
# Consolidated Flask-based requirements

# Flask and Extensions
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
flask-jwt-extended==4.5.2
flask-session==0.5.0

# Database
SQLAlchemy==2.0.21
psycopg2-binary==2.9.9
alembic==1.12.0

# Authentication & Security
PyJWT==2.8.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
werkzeug>=2.3.7
cryptography==41.0.4

# Google Ads API
google-ads==22.1.0
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# AI Integration
openai==1.3.0
httpx==0.24.1

# Async Support
aiohttp==3.8.5

# Email
flask-mail==0.9.1

# Additional Utilities
click==8.1.7
pyyaml==6.0.1

# Data Processing
pandas==2.0.3
numpy==1.25.1
scikit-learn==1.3.0
scipy==1.11.1

# Utilities
redis==5.0.1
hiredis==2.2.3
celery==5.3.1
requests==2.31.0
pydantic==2.0.3
email-validator==2.0.0
python-multipart==0.0.6
pytz==2023.3

# Development
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.9.1
flake8==6.1.0

# API Documentation
flasgger==0.9.7.1

# Monitoring & Logging
psutil==5.9.5
python-json-logger==2.0.7

# Production Server
gunicorn==21.2.0
eventlet==0.33.3